requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.12.0
redis==3.4.1
vk-api==11.8.0
